WIP: Continuing working on switching to asyncs, and now switching the Logger to use a different quiescence-detection mode.
In particular, this commit dumps the [known-unsafe] WaitTids method.
This commit also drops old code for forking workers with forkWithExceptions,
which was now confirmed to cause major problems in the testing framework.

With the loss of WaitTids, it was possible to push through a refactor:
  * logger field is no longer maybe
  * initLogger gets folded into Sched.new

Unfortunately, closing the logger hangs with the new WaitNum method.  It needs
to be debugged; it's now spinning in a backoff loop.  Warning: this WIP commit
includes a bunch of junk debugging messages marked with TEMP:.